refactor: Break down page into multiple components.
feat: Add syntax highlighting
