+++ types/react-syntax-highlighter.d.ts
@@ -1,0 +1,2 @@
+declare module 'react-syntax-highlighter';
+declare module 'react-syntax-highlighter/*';
