+++ types/remark-gfm.d.ts
@@ -1,0 +1,4 @@
+declare module 'remark-gfm' {
+  const plugin: any;
+  export default plugin;
+}
